% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scm_fit.R
\name{plot_scm_fit}
\alias{plot_scm_fit}
\title{Plot Donor Pool Optimization Fit}
\usage{
plot_scm_fit(scm_output, line_color = "black", dot_color = "black")
}
\arguments{
\item{scm_output}{Output from \code{prep_scm_panel()}.}

\item{line_color}{Line color.}

\item{dot_color}{Point color.}
}
\value{
A ggplot object.
}
\description{
Plots the summed normalized mean squared error used to select the optimal
donor pool size.
}
