% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelSCM-package.R
\docType{package}
\name{panelSCM-package}
\alias{panelSCM-package}
\alias{panelSCM}
\title{panelSCM: Synthetic Control with Heterogeneous Treatment Effects}
\description{
Implements a panel-data extension of the Synthetic Control Method to estimate
heterogeneous treatment effects. The methodology is described in Baumert and
Harmening (forthcoming).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mharmening/panelSCM}
  \item Report bugs at \url{https://github.com/mharmening/panelSCM/issues}
}

}
\author{
\strong{Maintainer}: Morten Harmening \email{m.harmening@ipw.uni-hannover.de}

Authors:
\itemize{
  \item Jona-Frederik Baumert
}

}
\keyword{internal}
