% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelSCM-package.R
\docType{package}
\name{panelSCM-package}
\alias{panelSCM}
\alias{panelSCM-package}
\title{panelSCM: What the Package Does (One Line, Title Case)}
\description{
Implements a panel-data extension of the Synthetic Control Method to estimate heterogeneous treatment effects. The approach constructs unit-specific synthetic controls, aligns individual treatment effects on a relative time scale, and estimates heterogeneous average treatment effects using kernel density estimation and highest density regions. The methodology is described in Baumert and Harmening (forthcoming).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
